fix(typping): fix type declaration
