+++ @types/yggdrasil/service/types/route.d.ts
@@ -1,9 +1,20 @@
 /// <reference types="node" />
 import { RateLimitPluginOptions } from "@fastify/rate-limit";
 import { TypeBoxTypeProvider } from "@fastify/type-provider-typebox";
+import boom from "@hapi/boom";
 import { Static, TSchema } from "@sinclair/typebox";
 import { ContextConfigDefault, FastifyInstance, FastifyPluginCallback, FastifySchema, onRequestHookHandler, preHandlerAsyncHookHandler, preHandlerHookHandler, RouteHandlerMethod } from "fastify";
 import { IncomingMessage, Server, ServerResponse } from "http";
+import { MulterFile } from "../../utils";
+declare module "fastify" {
+    interface FastifyRequest {
+        file?: MulterFile;
+        files?: MulterFile[] | Record<string, MulterFile[]>;
+    }
+    interface FastifyReply {
+        boom: typeof boom;
+    }
+}
 export interface Schema {
     summary: string;
     description?: string;
