+++ package.json
@@ -85,7 +85,7 @@
     "@types/ioredis": "^4.28.10",
     "@types/memoizee": "^0.4.8",
     "@types/papaparse": "^5.3.5",
-    "jormungand": "https://github.com/belo-app/jormungand.git#0.2.4"
+    "jormungand": "https://github.com/belo-app/jormungand.git#0.3.0"
   },
   "release-it": {
     "git": {
