--- src/service/environment.d.ts
@@ -1,14 +1,0 @@
-import boom from "@hapi/boom";
-
-import { MulterFile } from "../utils";
-
-declare module "fastify" {
-  interface FastifyRequest {
-    file?: MulterFile;
-    files?: MulterFile[] | Record<string, MulterFile[]>;
-  }
-
-  interface FastifyReply {
-    boom: typeof boom;
-  }
-}
